refactor: Rename `lib/resolveParams.js`
