+++ lib/param-command.js
@@ -3,7 +3,7 @@
 const { isEmpty } = require('lodash');
 const { writeText, style } = require('@serverless/utils/log');
 const log = require('./log');
-const resolveParams = require('./resolveParams');
+const resolveParams = require('./resolve-params');
 
 module.exports = {
   get: async (context) => {
