+++ lib/variables.js
@@ -3,7 +3,7 @@
 const _ = require('lodash');
 const ensureString = require('type/string/ensure');
 const resolveOutput = require('./resolve-output');
-const resolveParams = require('./resolveParams');
+const resolveParams = require('./resolve-params');
 const throwAuthError = require('./throwAuthError');
 
 // functions for new way of getting variables
